Fill out mass, radius, color for bodies
